+++ apps/remnant2toolkit/app/(builds)/_constants/max-trait-amount.ts
@@ -3,4 +3,3 @@
  * be added to the builder.
  */
 export const MAX_TRAIT_AMOUNT = 110;
-export const TRAIT_TRAITOR_BONUS = 30;